w3c-reports/12092014/mac-wk2/runned-tests/2dcontext/compositing/2d.composite.uncovered.image.copy-expected.txt

2d.composite.uncovered.image.copy
drawImage() draws pixels not covered by the source object as (0,0,0,0), and does not leave the pixels unchanged.
Actual output:
Expected output:


FAIL drawImage() draws pixels not covered by the source object as (0,0,0,0), and does not leave the pixels unchanged. assert_approx_equals: Green channel of the pixel at (15, 15) expected 0 +/- 5 but got 255(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_approx_equals@http://localhost:8800/resources/testharness.js:425:15
_assertPixelApprox@http://localhost:8800/common/canvas-tests.js:48:25
http://localhost:8800/2dcontext/compositing/2d.composite.uncovered.image.copy.html:27:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:62:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:59:15)
